requirements.txt
asgiref==3.2.10
Django==3.1.1
pytz==2020.1
sqlparse==0.3.1
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
mysqlclient==2.0.1
pandas==1.1.2
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorflow==2.1.0
gensim==3.8.3
numpy~=1.19.2
plotly==4.11.0
funcsigs==1.0.2
implicit==0.4.4
scikit-surprise==1.1.1